// nest/src/interfaces/AnimeKodik.ts
export interface IAnimeKodik {
  shikimori_id: string;
  title: string;
  title_orig: string;
  description: string;
  genres: string[];
  tags: string[];
  rating: number;
  episodes: any;
  status: string;
  studio: string;
  posterUrl: string;
  playerUrl: string;
  screenshots: string[];
  year: string;
  kind: string;
  duration: string;
  countries: string;
  ageRating: string;
  minimalAge: string;
}